Generate FPC file

Compute percentages changes across time

Usage

ff_ppts_lrce_fpc(
  df_flr,
  ar_it_years_chg = c(1960, 1980, 2000, 2020),
  ls_chg_years = list(chg_80v60 = c(1960, 1980), chg_00v80 = c(1980, 2000), chg_20v00 =
    c(2000, 2020), chg_20v80 = c(1980, 2020)),
  verbose = FALSE
)

Arguments

df_flr

Dataframe input, output of PrjCompPPTS::ff_ppts_lrce_flr

ar_it_years_chg

Array of integer, years to select for computing changes

verbose

Boolean print details

ls_chg_year

List of Array each array with start and end date and list contains new variable names

Value

A dataset with ratios, levels, and percentage changes.

References

PrjThaiHFID-issue-7, Step 2

Author

Fan Wang, http://fanwangecon.github.io