Panel data on school age population, students, teachers, and schools

Annual data available for different time frames across countries.

Data on school age population and teachers available globally. Data on school age population, teachers and schools available for Western European and Eastern Asian countries.

Usage

ppts_easia_weuro_world

Format

A data frame with 9 rows and 5 variables:

Country

Country name

countrycode

Three letter country code

countrycode

Values equal to nationwide, urban and non-urban for Korea

stats_youthpop

Youth population age 0 to 15

stats_student

Number of students in primary school

stats_teacher

Number of teachers in primary school

stats_school

Number of primary schools

stats_gdp

Annual GDP

stats_enroll_ratio

Primary school enrollment ratio

Source

https://fanwangecon.github.io/

Details

Aggregate national for countries except in the case of Korea.

For Korea, have with-in country data aggregated to compare between metropolitan and non-metropolitan areas. Also has data available in Korea for all regions

Examples

data(ppts_easia_weuro_world)
print(ppts_easia_weuro_world)
#> # A tibble: 17,712 × 10
#>    location_name location_code location_level  year stats_youthpop stats_student
#>    <fct>         <fct>         <fct>          <dbl>          <dbl>         <dbl>
#>  1 Aruba         ABW           country         1960          23769            NA
#>  2 Aruba         ABW           country         1961          24035            NA
#>  3 Aruba         ABW           country         1962          24139            NA
#>  4 Aruba         ABW           country         1963          24091            NA
#>  5 Aruba         ABW           country         1964          23924            NA
#>  6 Aruba         ABW           country         1965          23683            NA
#>  7 Aruba         ABW           country         1966          23479            NA
#>  8 Aruba         ABW           country         1967          23160            NA
#>  9 Aruba         ABW           country         1968          22773            NA
#> 10 Aruba         ABW           country         1969          22398            NA
#> # ℹ 17,702 more rows
#> # ℹ 4 more variables: stats_teacher <dbl>, stats_school <dbl>, stats_gdp <dbl>,
#> #   stats_enroll_ratio <dbl>