Elasticities across all potential variable pairs and for different time frames, non-interpolated

Compute elasticities for different numerator and denominator combinations, using non-interpolated data.

Usage

ppts_easia_weuro_world_elas_interp1

Format

A data frame:

location_code

Short string code for country and regions

location_level

Country, or regional or sub-country

variable

Statistical variable

year_bins_type

The type of year binning

year_bins

Year bins given year binning

pchg_interp1

Percentage change from start to end of bin associated with variable (for the denominator), with interpolation and extrapolation

value_interp1

Level values for variables, with interpolation and extrapolation

variable_numerator

The variable, whose percentage change is computed, in the numerator

pchg_interp1_numerator

pchg but for the variable_numerator variable, with interpolation and extrapolation

elasticity_interp1

pchg_interp1_numerator over pchg_interp1 to compute elasticities over different windows for all possible variable variable_numerator and variable combinations.

Source

World bank, national statistics websites of W. Europe and E. Asia countries

Examples

data(ppts_easia_weuro_world_elas_interp1)
summary(ppts_easia_weuro_world_elas_interp1)
#>  location_code          location_level      variable          year_bins_type 
#>  AUT    :   915   country      :96940   gdp     :29475   1920t2020i05:60015  
#>  CHN    :   780   multicountry :24760   school  : 3430   1920t2020i10:31805  
#>  JPN    :   780   multiprovince:  750   student :31845   1920t2020i20:17180  
#>  CHE    :   690   province     : 6375   teacher :30290   1925t2015i15:19825  
#>  KOR    :   625                         youthpop:33785   1940t2020i01:    0  
#>  BGD    :   530                                                              
#>  (Other):124505                                                              
#>   year_bins          pchg_interp1      value_interp1       variable_numerator
#>  Length:128825      Min.   :  -0.873   Min.   :2.200e+01   Length:128825     
#>  Class :character   1st Qu.:  -0.009   1st Qu.:6.844e+03   Class :character  
#>  Mode  :character   Median :   0.105   Median :1.120e+05   Mode  :character  
#>                     Mean   :   0.305   Mean   :2.617e+07                     
#>                     3rd Qu.:   0.278   3rd Qu.:2.234e+06                     
#>                     Max.   :1663.739   Max.   :1.977e+09                     
#>                     NA's   :20750                                            
#>  pchg_interp1_numerator elasticity_interp1
#>  Min.   :  -0.87        Min.   :-Inf      
#>  1st Qu.:  -0.01        1st Qu.:0.07      
#>  Median :   0.11        Median :1.00      
#>  Mean   :   0.31        Mean   : NaN      
#>  3rd Qu.:   0.28        3rd Qu.:1.00      
#>  Max.   :1663.74        Max.   : Inf      
#>  NA's   :47082          NA's   :50369     
print(ppts_easia_weuro_world_elas_interp1)
#> # A tibble: 128,825 × 10
#>    location_code location_level variable year_bins_type year_bins pchg_interp1
#>    <fct>         <fct>          <fct>    <fct>          <chr>            <dbl>
#>  1 ABW           country        gdp      1920t2020i05   1981-1985        1.25 
#>  2 ABW           country        gdp      1920t2020i05   1981-1985        1.25 
#>  3 ABW           country        gdp      1920t2020i05   1981-1985        1.25 
#>  4 ABW           country        gdp      1920t2020i05   1981-1985        1.25 
#>  5 ABW           country        gdp      1920t2020i05   1981-1985        1.25 
#>  6 ABW           country        gdp      1920t2020i05   1986-1990        0.903
#>  7 ABW           country        gdp      1920t2020i05   1986-1990        0.903
#>  8 ABW           country        gdp      1920t2020i05   1986-1990        0.903
#>  9 ABW           country        gdp      1920t2020i05   1986-1990        0.903
#> 10 ABW           country        gdp      1920t2020i05   1986-1990        0.903
#> # ℹ 128,815 more rows
#> # ℹ 4 more variables: value_interp1 <dbl>, variable_numerator <chr>,
#> #   pchg_interp1_numerator <dbl>, elasticity_interp1 <dbl>