Interpolated and extrapolated levels and changes panel across time bins

Panel data on school age population, students, teachers, and schools, levels and changes, interpolated and extrapolated

Usage

ppts_easia_weuro_world_pchg

Format

A data frame with 266 rows and 7 variables:

location_code

Short string code for country and regions

location_level

Country, or regional or sub-country

variable

Statistical variable

year_bins_type

The type of year binning, 1 year, 10 years, 20 years, etc.

year_bins

Year bins given year binning, labels of different years, decades, bi-decades, etc.

value

Level at start to end of bin

value_interp1

value with interpolation and extrapolation

pchg

Percentage change from start to end of bin

pchg_interp1

pchg with interpolation and extrapolation

Source

articles/ffv_gen_percent_changes, See online appendix C of Hannum, Kim, and Wang (2024).

Examples

data(ppts_easia_weuro_world_pchg)
summary(ppts_easia_weuro_world_pchg)
#>  location_code         location_level      variable          year_bins_type 
#>  AUT    :  609   country      :62172   gdp     :18763   1920t2020i05:12003  
#>  JPN    :  507   multicountry :15896   school  : 2222   1920t2020i10: 6361  
#>  CHN    :  504   multiprovince:  486   student :20673   1920t2020i20: 3436  
#>  CHE    :  463   province     : 4131   teacher :19549   1925t2015i15: 3965  
#>  KOR    :  405                         youthpop:21478   1940t2020i01:56920  
#>  BGD    :  340                                                              
#>  (Other):79857                                                              
#>   year_bins              pchg         pchg_interp1          value          
#>  Length:82685       Min.   :-0.715   Min.   :  -0.873   Min.   :2.200e+01  
#>  Class :character   1st Qu.:-0.004   1st Qu.:  -0.004   1st Qu.:7.412e+03  
#>  Mode  :character   Median : 0.018   Median :   0.020   Median :1.505e+05  
#>                     Mean   : 0.055   Mean   :   0.097   Mean   :2.921e+07  
#>                     3rd Qu.: 0.050   3rd Qu.:   0.057   3rd Qu.:2.729e+06  
#>                     Max.   : 9.758   Max.   :1663.739   Max.   :1.977e+09  
#>                     NA's   :19120    NA's   :4155       NA's   :11215      
#>  value_interp1      
#>  Min.   :2.000e+00  
#>  1st Qu.:6.790e+03  
#>  Median :1.125e+05  
#>  Mean   :2.600e+07  
#>  3rd Qu.:2.213e+06  
#>  Max.   :1.977e+09  
#>                     
print(ppts_easia_weuro_world_pchg)
#> # A tibble: 82,685 × 9
#>    location_code location_level variable year_bins_type year_bins    pchg
#>    <fct>         <fct>          <fct>    <fct>          <chr>       <dbl>
#>  1 ABW           country        gdp      1920t2020i05   1981-1985 NA     
#>  2 ABW           country        gdp      1920t2020i05   1986-1990 NA     
#>  3 ABW           country        gdp      1920t2020i05   1991-1995  0.0532
#>  4 ABW           country        gdp      1920t2020i05   1996-2000  0.0640
#>  5 ABW           country        gdp      1920t2020i05   2001-2005 -0.0504
#>  6 ABW           country        gdp      1920t2020i05   2006-2010 -0.129 
#>  7 ABW           country        gdp      1920t2020i05   2011-2015  0.0982
#>  8 ABW           country        gdp      1920t2020i05   2016-2020 NA     
#>  9 ABW           country        student  1920t2020i05   1991-1995 NA     
#> 10 ABW           country        student  1920t2020i05   1996-2000 NA     
#> # ℹ 82,675 more rows
#> # ℹ 3 more variables: pchg_interp1 <dbl>, value <dbl>, value_interp1 <dbl>