Panel data on school age population, students, teachers, and schools

Annual data available for different time frames across countries.

Data on school age population and teachers available globally. Data on school age population, teachers and schools available for Western European and Eastern Asian countries.

Usage

ppts_easia_weuro_world

Format

A data frame with 9 rows and 5 variables:

Country

Country name

countrycode

Three letter country code

countrycode

Values equal to nationwide, urban and non-urban for Korea

stats_youthpop

Youth population age 0 to 15

stats_student

Number of students in primary school

stats_teacher

Number of teachers in primary school

stats_school

Number of primary schools

stats_gdp

Annual GDP

stats_enroll_ratio

Primary school enrollment ratio

Source

See online appendix B of Hannum, Kim, and Wang (2024).

Details

Aggregate national for countries except in the case of Korea.

For Korea, within country data aggregated to compare between metropolitan and non-metropolitan areas as well as province-specific data.

Examples

data(ppts_easia_weuro_world)
summary(ppts_easia_weuro_world)
#>                      location_name   location_code         location_level 
#>  Switzerland                :  157   CHE    :  157   country      :13529  
#>  Austria                    :   98   AUT    :   98   multicountry : 3100  
#>  Japan                      :   75   JPN    :   75   multiprovince:  114  
#>  China                      :   73   CHN    :   73   province     :  969  
#>  Afghanistan                :   62   ABW    :   62                        
#>  Africa Eastern and Southern:   62   AFE    :   62                        
#>  (Other)                    :17185   (Other):17185                        
#>       year      stats_youthpop      stats_student       stats_teacher     
#>  Min.   :1864   Min.   :1.272e+04   Min.   :     1008   Min.   :      40  
#>  1st Qu.:1975   1st Qu.:7.113e+05   1st Qu.:   180278   1st Qu.:    7402  
#>  Median :1991   Median :2.704e+06   Median :   814673   Median :   35938  
#>  Mean   :1990   Mean   :7.559e+07   Mean   : 28134713   Mean   : 1210151  
#>  3rd Qu.:2006   3rd Qu.:1.745e+07   3rd Qu.:  5809414   3rd Qu.:  340705  
#>  Max.   :2022   Max.   :1.977e+09   Max.   :742831104   Max.   :32652932  
#>                 NA's   :3067        NA's   :6165        NA's   :7885      
#>   stats_school         stats_gdp      stats_enroll_ratio
#>  Min.   :     22.0   Min.   :   144   Min.   :  3.00    
#>  1st Qu.:    184.2   1st Qu.:  1277   1st Qu.: 93.00    
#>  Median :    559.0   Median :  3424   Median :101.00    
#>  Mean   :  34613.1   Mean   : 10500   Mean   : 97.21    
#>  3rd Qu.:   3521.5   3rd Qu.: 11308   3rd Qu.:107.00    
#>  Max.   :1681939.0   Max.   :181709   Max.   :222.00    
#>  NA's   :16190       NA's   :5585     NA's   :7392      
print(ppts_easia_weuro_world)
#> # A tibble: 17,712 × 10
#>    location_name location_code location_level  year stats_youthpop stats_student
#>    <fct>         <fct>         <fct>          <dbl>          <dbl>         <dbl>
#>  1 Aruba         ABW           country         1960          23769            NA
#>  2 Aruba         ABW           country         1961          24035            NA
#>  3 Aruba         ABW           country         1962          24139            NA
#>  4 Aruba         ABW           country         1963          24091            NA
#>  5 Aruba         ABW           country         1964          23924            NA
#>  6 Aruba         ABW           country         1965          23683            NA
#>  7 Aruba         ABW           country         1966          23479            NA
#>  8 Aruba         ABW           country         1967          23160            NA
#>  9 Aruba         ABW           country         1968          22773            NA
#> 10 Aruba         ABW           country         1969          22398            NA
#> # ℹ 17,702 more rows
#> # ℹ 4 more variables: stats_teacher <dbl>, stats_school <dbl>, stats_gdp <dbl>,
#> #   stats_enroll_ratio <dbl>